// src/api/generated/schemas/ticketSummary.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * WorkLog API
 * API para gerenciamento de tickets e interações de suporte ao cliente. Autenticação via cookie HttpOnly `worklog_access` emitido pelo endpoint `POST /worklog/auth/login` (o browser anexa o cookie automaticamente em chamadas subsequentes).
 * OpenAPI spec version: v1.0.0
 */
import type { ClientSummary } from './clientSummary';
import type { SystemResponse } from './systemResponse';
import type { TicketSummaryPriority } from './ticketSummaryPriority';
import type { TicketSummaryStatus } from './ticketSummaryStatus';
import type { UserSummary } from './userSummary';

export interface TicketSummary {
  publicId?: string;
  title?: string;
  description?: string;
  status?: TicketSummaryStatus;
  priority?: TicketSummaryPriority;
  createdAt?: string;
  updatedAt?: string;
  completedAt?: string;
  client?: ClientSummary;
  system?: SystemResponse;
  user?: UserSummary;
}